prereqs: use prebuilt `z3` from the `z3-solver` Python package
`z3` takes forever to build, but `z3` publishes prebuilt Python wheels,
so install that (`z3-solver`) instead.
Then do a normal `cmake` build of `z3`, except for the main build step,
just copy `libz3.so` and `z3` from `z3-solver`.
We still need to do the `cmake` installation to ensure
all of the headers and config files end up in the right place.